backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
pydantic-settings>=2.0.0
fastapi-cors
pyautogen==0.7.6  # For agent functionality 